Refactor git operations
