+++ app/src/main/java/com/zeapo/pwdstore/git/BaseGitActivity.kt
@@ -20,6 +20,13 @@ import com.zeapo.pwdstore.R
 import com.zeapo.pwdstore.git.config.ConnectionMode
 import com.zeapo.pwdstore.git.config.Protocol
 import com.zeapo.pwdstore.git.config.SshApiSessionFactory
+import com.zeapo.pwdstore.git.operation.BreakOutOfDetached
+import com.zeapo.pwdstore.git.operation.CloneOperation
+import com.zeapo.pwdstore.git.operation.GitOperation
+import com.zeapo.pwdstore.git.operation.PullOperation
+import com.zeapo.pwdstore.git.operation.PushOperation
+import com.zeapo.pwdstore.git.operation.ResetToRemoteOperation
+import com.zeapo.pwdstore.git.operation.SyncOperation
 import com.zeapo.pwdstore.utils.PasswordRepository
 import com.zeapo.pwdstore.utils.PreferenceKeys
 import com.zeapo.pwdstore.utils.getEncryptedPrefs
@@ -190,12 +197,12 @@ abstract class BaseGitActivity : AppCompatActivity() {
 
             val localDir = requireNotNull(PasswordRepository.getRepositoryDirectory(this))
             val op = when (operation) {
-                REQUEST_CLONE, GitOperation.GET_SSH_KEY_FROM_CLONE -> CloneOperation(localDir, this).setCommand(url!!)
-                REQUEST_PULL -> PullOperation(localDir, this).setCommand()
-                REQUEST_PUSH -> PushOperation(localDir, this).setCommand()
-                REQUEST_SYNC -> SyncOperation(localDir, this).setCommands()
-                BREAK_OUT_OF_DETACHED -> BreakOutOfDetached(localDir, this).setCommands()
-                REQUEST_RESET -> ResetToRemoteOperation(localDir, this).setCommands()
+                REQUEST_CLONE, GitOperation.GET_SSH_KEY_FROM_CLONE -> CloneOperation(localDir, url!!, this)
+                REQUEST_PULL -> PullOperation(localDir, this)
+                REQUEST_PUSH -> PushOperation(localDir, this)
+                REQUEST_SYNC -> SyncOperation(localDir, this)
+                BREAK_OUT_OF_DETACHED -> BreakOutOfDetached(localDir, this)
+                REQUEST_RESET -> ResetToRemoteOperation(localDir, this)
                 SshApiSessionFactory.POST_SIGNATURE -> return
                 else -> {
                     tag(TAG).e { "Operation not recognized : $operation" }
