+++ app/src/main/java/com/zeapo/pwdstore/git/operation/CloneOperation.kt
@@ -2,41 +2,33 @@
  * Copyright © 2014-2020 The Android Password Store Authors. All Rights Reserved.
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package com.zeapo.pwdstore.git
+package com.zeapo.pwdstore.git.operation
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.zeapo.pwdstore.R
+import com.zeapo.pwdstore.git.GitAsyncTask
 import java.io.File
-import org.eclipse.jgit.api.CloneCommand
 import org.eclipse.jgit.api.Git
+import org.eclipse.jgit.api.GitCommand
 
 /**
  * Creates a new clone operation
  *
  * @param fileDir the git working tree directory
+ * @param uri URL to clone the repository from
  * @param callingActivity the calling activity
  */
-class CloneOperation(fileDir: File, callingActivity: AppCompatActivity) : GitOperation(fileDir, callingActivity) {
+class CloneOperation(fileDir: File, uri: String, callingActivity: AppCompatActivity) : GitOperation(fileDir, callingActivity) {
 
-    /**
-     * Sets the command using the repository uri
-     *
-     * @param uri the uri of the repository
-     * @return the current object
-     */
-    fun setCommand(uri: String): CloneOperation {
-        this.command = Git.cloneRepository()
-            .setCloneAllBranches(true)
-            .setDirectory(repository?.workTree)
-            .setURI(uri)
-        return this
-    }
+    override val commands: Array<GitCommand<out Any>> = arrayOf(
+        Git.cloneRepository().setCloneAllBranches(true).setDirectory(repository?.workTree).setURI(uri),
+    )
 
     override fun execute() {
-        (this.command as? CloneCommand)?.setCredentialsProvider(this.provider)
-        GitAsyncTask(callingActivity, this, Intent()).execute(this.command)
+        setCredentialProvider()
+        GitAsyncTask(callingActivity, this, Intent()).execute(*commands)
     }
 
     override fun onError(err: Exception) {
