+++ app/src/main/java/com/zeapo/pwdstore/git/operation/PushOperation.kt
@@ -2,15 +2,16 @@
  * Copyright © 2014-2020 The Android Password Store Authors. All Rights Reserved.
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package com.zeapo.pwdstore.git
+package com.zeapo.pwdstore.git.operation
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.zeapo.pwdstore.R
+import com.zeapo.pwdstore.git.GitAsyncTask
 import java.io.File
 import org.eclipse.jgit.api.Git
-import org.eclipse.jgit.api.PullCommand
+import org.eclipse.jgit.api.GitCommand
 
 /**
  * Creates a new git operation
@@ -18,34 +19,23 @@ import org.eclipse.jgit.api.PullCommand
  * @param fileDir the git working tree directory
  * @param callingActivity the calling activity
  */
-class PullOperation(fileDir: File, callingActivity: AppCompatActivity) : GitOperation(fileDir, callingActivity) {
+class PushOperation(fileDir: File, callingActivity: AppCompatActivity) : GitOperation(fileDir, callingActivity) {
 
-    /**
-     * Sets the command
-     *
-     * @return the current object
-     */
-    fun setCommand(): PullOperation {
-        this.command = Git(repository)
-            .pull()
-            .setRebase(true)
-            .setRemote("origin")
-        return this
-    }
+    override val commands: Array<GitCommand<out Any>> = arrayOf(
+        Git(repository).push().setPushAll().setRemote("origin"),
+    )
 
     override fun execute() {
-        (this.command as? PullCommand)?.setCredentialsProvider(this.provider)
-        GitAsyncTask(callingActivity, this, Intent()).execute(this.command)
+        setCredentialProvider()
+        GitAsyncTask(callingActivity, this, Intent()).execute(*commands)
     }
 
     override fun onError(err: Exception) {
+        // TODO handle the "Nothing to push" case
         super.onError(err)
         MaterialAlertDialogBuilder(callingActivity)
             .setTitle(callingActivity.resources.getString(R.string.jgit_error_dialog_title))
-            .setMessage("Error occurred during the pull operation, " +
-                callingActivity.resources.getString(R.string.jgit_error_dialog_text) +
-                err.message +
-                "\nPlease check the FAQ for possible reasons why this error might occur.")
+            .setMessage(callingActivity.getString(R.string.jgit_error_push_dialog_text) + err.message)
             .setPositiveButton(callingActivity.resources.getString(R.string.dialog_ok)) { _, _ -> callingActivity.finish() }
             .show()
     }
