+++ app/src/main/java/com/zeapo/pwdstore/git/operation/ResetToRemoteOperation.kt
@@ -2,19 +2,17 @@
  * Copyright © 2014-2020 The Android Password Store Authors. All Rights Reserved.
  * SPDX-License-Identifier: GPL-3.0-only
  */
-package com.zeapo.pwdstore.git
+package com.zeapo.pwdstore.git.operation
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import androidx.preference.PreferenceManager
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.zeapo.pwdstore.R
+import com.zeapo.pwdstore.git.GitAsyncTask
 import com.zeapo.pwdstore.utils.PreferenceKeys
 import java.io.File
-import org.eclipse.jgit.api.Git
-import org.eclipse.jgit.api.GitCommand
 import org.eclipse.jgit.api.ResetCommand
-import org.eclipse.jgit.api.TransportCommand
 
 /**
  * Creates a new git operation
@@ -24,35 +22,20 @@ import org.eclipse.jgit.api.TransportCommand
  */
 class ResetToRemoteOperation(fileDir: File, callingActivity: AppCompatActivity) : GitOperation(fileDir, callingActivity) {
 
-    private lateinit var commands: List<GitCommand<out Any>>
+    private val remoteBranch = PreferenceManager
+        .getDefaultSharedPreferences(callingActivity.applicationContext)
+        .getString(PreferenceKeys.GIT_BRANCH_NAME, "master")
 
-    /**
-     * Sets the command
-     *
-     * @return the current object
-     */
-    fun setCommands(): ResetToRemoteOperation {
-        val remoteBranch = PreferenceManager
-            .getDefaultSharedPreferences(callingActivity.applicationContext)
-            .getString(PreferenceKeys.GIT_BRANCH_NAME, "master")
-        val git = Git(repository)
-        val cmds = arrayListOf(
-            git.add().addFilepattern("."),
-            git.fetch().setRemote("origin"),
-            git.reset().setRef("origin/$remoteBranch").setMode(ResetCommand.ResetType.HARD)
-        )
-        if (git.branchList().call().none { it.name == remoteBranch }) {
-            cmds.add(
-                git.branchCreate().setName(remoteBranch).setForce(true)
-            )
-        }
-        commands = cmds
-        return this
-    }
+    override val commands = arrayOf(
+        git.add().addFilepattern("."),
+        git.fetch().setRemote("origin"),
+        git.reset().setRef("origin/$remoteBranch").setMode(ResetCommand.ResetType.HARD),
+        git.branchCreate().setName(remoteBranch).setForce(false),
+    )
 
     override fun execute() {
-        commands.filterIsInstance<TransportCommand<*, *>>().map { it.setCredentialsProvider(provider) }
-        GitAsyncTask(callingActivity, this, Intent()).execute(*commands.toTypedArray())
+        setCredentialProvider()
+        GitAsyncTask(callingActivity, this, Intent()).execute(*commands)
     }
 
     override fun onError(err: Exception) {
